Break out duplicated code from french providers into a new base provider class Select magnets instead of torrent file unless using black hole. With black hole, try to use torrent file first. Enable rss for cpasbien and torrent911 Drop html5lib parser explicitly being set for providers, let the BS4Parser try to picl the most efficient parser
