+++ sickchill/oldbeard/providers/abnormal.py
@@ -87,7 +87,7 @@ class Provider(TorrentProvider):
                 if not data:
                     continue
 
-                with BS4Parser(data, "html5lib") as html:
+                with BS4Parser(data) as html:
                     torrent_table = html.find(class_="table-rows")
                     torrent_rows = torrent_table("tr") if torrent_table else []
