+++ sickchill/oldbeard/providers/archetorrent.py
@@ -101,7 +101,7 @@ class Provider(TorrentProvider):
                 if not data:
                     continue
 
-                with BS4Parser(data, "html5lib") as html:
+                with BS4Parser(data) as html:
                     torrent_table = html.find(class_="ttable_headinner")
                     torrent_rows = torrent_table("tr") if torrent_table else []
