+++ sickchill/oldbeard/providers/bjshare.py
@@ -137,7 +137,7 @@ class Provider(TorrentProvider):
             return ret
 
         items = []
-        with BS4Parser(data, "html5lib") as html:
+        with BS4Parser(data) as html:
             torrent_table = html.find("table", id="torrent_table")
             torrent_rows = torrent_table("tr") if torrent_table else []
