+++ sickchill/oldbeard/providers/demonoid.py
@@ -60,7 +60,7 @@ class Provider(TorrentProvider):
                     logger.debug("No data returned from provider")
                     continue
 
-                with BS4Parser(data, "html5lib") as html:
+                with BS4Parser(data) as html:
                     for result in html("img", alt="Download torrent"):
                         try:
                             data_row = result.find_parent("tr")
