+++ sickchill/oldbeard/providers/hdtorrents.py
@@ -99,7 +99,7 @@ class Provider(TorrentProvider):
 
                 data = data[index:]
 
-                with BS4Parser(data, "html5lib") as html:
+                with BS4Parser(data) as html:
                     if not html:
                         logger.debug("No html data parsed from provider")
                         continue
