+++ sickchill/oldbeard/providers/newpct.py
@@ -62,7 +62,7 @@ class Provider(TorrentProvider):
                 if not data:
                     continue
 
-                with BS4Parser(data, "html5lib") as html:
+                with BS4Parser(data) as html:
                     torrent_table = html.find("table", id="categoryTable")
                     torrent_rows = torrent_table("tr") if torrent_table else []
