+++ sickchill/oldbeard/providers/newznab.py
@@ -165,7 +165,7 @@ class NewznabProvider(NZBProvider):
             logger.warning(error_string)
             return False, return_categories, error_string
 
-        with BS4Parser(data, "html5lib") as html:
+        with BS4Parser(data) as html:
             try:
                 self.torznab = html.find("server").get("title") == "Jackett"
             except AttributeError:
@@ -334,7 +334,7 @@ class NewznabProvider(NZBProvider):
                 if not data:
                     break
 
-                with BS4Parser(data, "html5lib") as html:
+                with BS4Parser(data) as html:
                     if not self._check_auth_from_data(html):
                         break
