+++ sickchill/oldbeard/providers/tntvillage.py
@@ -290,7 +290,7 @@ class Provider(TorrentProvider):
                         continue
 
                     try:
-                        with BS4Parser(data, "html5lib") as html:
+                        with BS4Parser(data) as html:
                             torrent_table = html.find("table", class_="copyright")
                             torrent_rows = torrent_table("tr") if torrent_table else []
