+++ sickchill/oldbeard/providers/torrent911.py
@@ -1,82 +1,14 @@
 import re
-from urllib.parse import urljoin
-
-import validators
 
 from sickchill import logger
 from sickchill.helper.common import convert_size, try_int
-from sickchill.oldbeard import tvcache
 from sickchill.oldbeard.bs4_parser import BS4Parser
-from sickchill.oldbeard.show_name_helpers import allPossibleShowNames
-from sickchill.providers.torrent.TorrentProvider import TorrentProvider
+from sickchill.providers.torrent.FrenchProvider import FrenchTorrentProvider
 
 
-class Provider(TorrentProvider):
+class Provider(FrenchTorrentProvider):
     def __init__(self):
-
-        super().__init__("Torrent911")
-
-        self.public = True
-        self.minseed = 0
-        self.minleech = 0
-        self._original_url = "https://www.torrent911.cc"
-        self._custom_url = None
-        self._used_url = None
-        self._recheck_url = True
-
-        self.proper_strings = ["PROPER", "REPACK"]
-        self.cache = tvcache.TVCache(self)
-        self.ability_status = self.PROVIDER_BACKLOG
-
-    def _retrieve_dllink_from_url(self, inner_url):
-        data = self.get_url(urljoin(self.url, inner_url), returns="text")
-        regex = r".*?function\s+redirect\(\).+?= '([^']+)'"
-        with BS4Parser(data, "html5lib") as html:
-            scripts = html.head.findAll("script")
-            if len(scripts):
-                script = scripts[-1].get("src")
-                matches = re.match(regex, script, re.S)
-                if matches:
-                    return urljoin(self.url, matches[1])
-
-            # try href
-            download_btns = html.findAll("div", {"class": "btn-download"})
-            for btn in download_btns:
-                link = btn.find("a").get("href")
-                if link.startswith("javascript"):
-                    return ""
-                if link.startswith("magnet"):
-                    continue
-                else:
-                    return urljoin(self.url, link)
-
-        return ""
-
-    def _get_custom_url(self):
-        return self._custom_url
-
-    def _set_custom_url(self, url):
-        if self._custom_url != url:
-            self._custom_url = url
-            self._recheck_url = True
-
-    def _get_provider_url(self):
-        if self._recheck_url:
-            if self.custom_url:
-                if validators.url(self.custom_url):
-                    self._used_url = self.custom_url
-                else:
-                    logger.warning("Invalid custom url set, please check your settings")
-
-            self._used_url = self._original_url
-
-        return self._used_url
-
-    def _set_provider_url(self, url):
-        self._used_url = url
-
-    url = property(_get_provider_url, _set_provider_url)
-    custom_url = property(_get_custom_url, _set_custom_url)
+        super().__init__("Torrent911", "https://www.torrent911.cc")
 
     def search(self, search_strings, age=0, ep_obj=None):
         results = []
@@ -89,13 +21,13 @@ class Provider(TorrentProvider):
 
                     search_url = self.url + "/recherche/" + search_string
                 else:
-                    search_url = self.url + "/torrents/series"
+                    search_url = self.url + "/torrents/series/date/desc"
 
                 data = self.get_url(search_url, returns="text")
                 if not data:
                     continue
 
-                with BS4Parser(data, "html5lib") as html:
+                with BS4Parser(data) as html:
                     # Skip column headers
                     for result in html.select("table.table-hover tr")[1:]:
                         try:
@@ -131,21 +63,3 @@ class Provider(TorrentProvider):
             results += items
 
         return results
-
-    def get_season_search_strings(self, episode):
-        search_string = {"Season": []}
-        for show_name in allPossibleShowNames(episode.show, season=episode.scene_season):
-            season = int(episode.scene_season)
-            if episode.show.air_by_date or episode.show.sports:
-                year = str(episode.airdate).split('-')[0]
-                season_string = f"{show_name} {year}"
-            elif episode.show.anime:
-                # use string below if you really want to search on season with number
-                # season_string = f"{show_name} Saison {season:0d}"
-                season_string = f"{show_name} Saison" # ignore season number to get all seasons in all formats
-            else:
-                season_string = f"{show_name} Saison {season:0d}"
-
-            search_string["Season"].append(season_string)
-
-        return [search_string]
