+++ sickchill/oldbeard/providers/tvchaosuk.py
@@ -77,7 +77,7 @@ class Provider(TorrentProvider):
                     logger.debug(_("No data returned from provider"))
                     continue
 
-                with BS4Parser(data, "html5lib") as html:
+                with BS4Parser(data) as html:
                     torrent_table = html.find(id="sortabletable")
                     torrent_rows = torrent_table("tr") if torrent_table else []
